# landing/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Q-Mol API",
    "description": "Molecular descriptor, similarity search, drug-likeness screen, and ADMET prediction API.\n\nGet a free key: `POST /signup`. Upgrade: https://qmol.app/checkout.html\n\nPostman collection: [qmol-postman.json](/qmol-postman.json)",
    "contact": {
      "name": "Q-Mol support",
      "email": "hi@qmol.app"
    },
    "license": {
      "name": "MIT (free tier) / Commercial (paid tiers)"
    },
    "version": "1.2.0"
  },
  "paths": {
...
    "/compute": {
      "post": {
        "summary": "Compute Free",
        "operationId": "compute_free_compute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/compute/premium": {
      "post": {
        "summary": "Compute Paid",
        "operationId": "compute_paid_compute_premium_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/generate": {
      "post": {
        "summary": "Generate Molecules",
        "description": "Generate novel SMILES using a character-level Markov model trained on a seed set of molecules. Returns valid, drug-like candidates.",
        "operationId": "generate_molecules_v1_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/optimize": {
      "post": {
        "summary": "Optimize Molecule",
        "description": "Optimize a single molecule toward a target property (logP, MW, TPSA, or QED) using a genetic algorithm.",
        "operationId": "optimize_molecule_v1_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/optimize/lead": {
      "post": {
        "summary": "Lead Optimization",
        "description": "Multi-objective lead optimization balancing multiple ADMET properties simultaneously (e.g., logP, MW, QED).",
        "operationId": "optimize_lead_v1_optimize_lead_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadOptimizeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/score": {
      "post": {
        "summary": "Score Molecule",
        "description": "Score a molecule against one or more ADMET / drug-likeness objectives. Returns desirability scores in [0, 1].",
        "operationId": "score_molecule_v1_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ComputeIn": {
        "properties": {
          "smiles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Smiles"
          }
        },
        "type": "object",
        "required": [
          "smiles"
        ],
        "title": "ComputeIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "GenerateIn": {
        "properties": {
          "seed_smiles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Seed Smiles"
          },
          "n": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "N",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "seed_smiles"
        ],
        "title": "GenerateIn"
      },
      "OptimizeIn": {
        "properties": {
          "smiles": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Smiles"
          },
          "target_property": {
            "type": "string",
            "pattern": "^(logP|MW|TPSA|QED)$",
            "title": "Target Property",
            "default": "logP"
          },
          "target_value": {
            "type": "number",
            "title": "Target Value"
          },
          "max_steps": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 50
          }
        },
        "type": "object",
        "required": [
          "smiles",
          "target_value"
        ],
        "title": "OptimizeIn"
      },
      "LeadOptimizeIn": {
        "properties": {
          "smiles": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Smiles"
          },
          "objectives": {
            "type": "object",
            "description": "Property: (target, tolerance)",
            "title": "Objectives"
          },
          "max_steps": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "smiles",
          "objectives"
        ],
        "title": "LeadOptimizeIn"
      },
      "ScoreIn": {
        "properties": {
          "smiles": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Smiles"
          },
          "objectives": {
            "type": "object",
            "description": "Property: target_value (uses default tolerances)",
            "title": "Objectives"
          },
          "weights": {
            "type": "object",
            "description": "Optional per-objective weights",
            "title": "Weights"
          }
        },
        "type": "object",
        "required": [
          "smiles"
        ],
        "title": "ScoreIn"
      }
    }
  }
}